% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{bp_loglik}
\alias{bp_loglik}
\title{bp_loglik}
\usage{
bp_loglik(model, params, init_pop, start_times, end_times, final_pop)
}
\arguments{
\item{model}{the \code{process_model} object representing the process generating the data}

\item{params}{the vector of parameters for which we are computing the likelihood}

\item{init_pop}{a \code{nobs x ntype} matrix with initial population for each observation}

\item{start_times}{the \code{nobs} length vector of times at which the initial populations were observed}

\item{end_times}{the \code{nobs} length vector of times at which the final populations were observed}

\item{final_pop}{the \code{nobs x mtype} matrix of final populations observed}
}
\value{
a dataframe with the moments vector for each timepoint
}
\description{
compute the log-likelihood of observing a set of data generated under a time-homogenous branching process model with some setting of parameters
}
