% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{branch}
\alias{branch}
\title{branch
Simulates a continuous-time time-inhomogenous markov branching process using the specified parameters. Uses C++ code for faster simulation.}
\usage{
branch(model, params, init_pop, time_obs, stops, reps, silent = FALSE,
  keep = FALSE, seed = NULL)
}
\arguments{
\item{model}{the \code{process_model} object representing the process being simulates}

\item{params}{the vector of parameters for which we are simulating the model}

\item{time_obs}{the vector of times at which to record the process state}

\item{stops}{\code{stop_list} object specifying stopping conditions for the system}

\item{reps}{the number of replicates to simulate}

\item{silent}{if true, verbose output will be shown.  Default: false}

\item{keep}{if true, the temporary comma-separated file generated while simulating while be kept.  if false, it will be deleted.  Default: false}

\item{seed}{seed for the random number generator.  If NULL, will use computer clock to set a random seed}
}
\description{
branch
Simulates a continuous-time time-inhomogenous markov branching process using the specified parameters. Uses C++ code for faster simulation.
}
