% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_valid}
\alias{check_valid}
\title{check_valid
 
helper for determining whether an expression is valid}
\usage{
check_valid(ast)
}
\description{
check_valid
 
helper for determining whether an expression is valid
}
