% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customize.R
\name{compile_timedep}
\alias{compile_timedep}
\title{compile_timedep}
\usage{
compile_timedep(cppfile)
}
\arguments{
\item{cppfile}{a character string giving the path name of a cpp file}
}
\description{
Allows the user to create a time-dependent rate function in C++ for use in
ESTIpop. Creates the appropriate header file and compiles the source file then
links into a shared object. The created shared object is used as a plugin.
}
\examples{
\dontrun{
compile_timedep(cppfile = "./plugin.cpp")
}
}
