% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{compute_mu_sigma}
\alias{compute_mu_sigma}
\title{compute_mu_sigma}
\usage{
compute_mu_sigma(model, params, start_time, end_time, init_pop)
}
\arguments{
\item{model}{the \code{process_model} object representing the process whose moments will be computed}

\item{params}{the vector of parameters to plug into the process model during moment computation}

\item{start_time}{the  start time of the moments to be computed}

\item{end_time}{the end time of the moments to be computed}
}
\value{
a list with the mean vector and covariance matrix
}
\description{
user-facing utility to compute the mean and covariance of a branching process population
}
