% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customize.R
\name{create_timedep_template}
\alias{create_timedep_template}
\title{Customization code for time-dependent rates
 taken with minor adaptions from:}
\usage{
create_timedep_template(cppfile = "custom_rate_plugin.cpp")
}
\arguments{
\item{cppfile}{file name of .cpp file to create}
}
\description{
McDonald, T. O., & Michor, F. (2017).
 SIApopr: a computational method to simulate evolutionary branching trees for
 analysis of tumor clonal evolution. Bioinformatics, 33(14), 2221-2223.
}
\details{
Original code author: T. O. McDonald (mcdonald(at)jimmy.harvard.edu)

create_timedep_template

Creates a time-dependent rate ".cpp" template file at the given location
which can be modified before running `compile_timedep`.
}
\examples{
\dontrun{
create_timedep_template(cppfile = "custom_rate_plugin.cpp")
}
}
