% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.pop}
\alias{.pop}
\title{.pop}
\usage{
.pop(string, pattern = ">", ...)
}
\arguments{
\item{string}{- character string where matches are sought.}

\item{pattern}{- character string to search for.}
}
\value{
.pop - string of everything after final pattern.
}
\description{
pops off last element of a string after final pattern in a vector
}
