% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.pop_off}
\alias{.pop_off}
\title{.pop_off}
\usage{
.pop_off(string, pattern = ">", ...)
}
\arguments{
\item{string}{- character string where matches are sought.}

\item{pattern}{- character string to search for.}
}
\value{
.pop_off - string with everything after final pattern removed.
}
\description{
pops off last instance of pattern in a string
}
