% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{estimate_td}
\usage{
estimate(model, init_pop, times, final_pop, initial, known = NULL,
  lower = NULL, upper = NULL, trace = 1)
}
\arguments{
\item{model}{the \code{process_model} object representing the process generating the data}

\item{init_pop}{a \code{nobs x ntype} matrix with initial population for each observation}

\item{final_pop}{the \code{nobs x mtype} matrix of final populations observed}

\item{initial}{vector of initial parameters estimates for MLE optimization}

\item{known}{boolean vector of known parameter rates, if NULL, all rates will be estimated}

\item{lower}{vector of lower bounds on rate parameters for optimization}

\item{upper}{vector of upper bounds on rate parameters for optimization}

\item{trace}{level of output for optimizer - see optim function, control - trace for "L-BFGS-B" method}

\item{params}{the vector of parameters for which we are computing the likelihood}

\item{time}{the \code{nobs} length vector containing the time between the initial and final population observations}
}
\description{
Estimates the rate parameters for a general multitype branching process with constant rates
}
