% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_td}
\alias{estimate_td}
\title{estimate_td}
\usage{
estimate_td(model, init_pop, start_times, end_times, final_pop, initial,
  known = NULL, lower = NULL, upper = NULL, trace = 1)
}
\arguments{
\item{model}{the \code{process_model} object representing the process generating the data}

\item{init_pop}{a \code{nobs x ntype} matrix with initial population for each observation}

\item{start_times}{the \code{nobs} length vector of times at which the initial populations were observed}

\item{end_times}{the \code{nobs} length vector of times at which the final populations were observed}

\item{final_pop}{the \code{nobs x mtype} matrix of final populations observed}

\item{initial}{vector of initial parameters estimates for MLE optimization}

\item{known}{boolean vector of known parameter rates, if NULL, all rates will be estimated}

\item{lower}{vector of lower bounds on rate parameters for optimization}

\item{upper}{vector of upper bounds on rate parameters for optimization}

\item{trace}{level of output for optimizer - see optim function, control - trace for "L-BFGS-B" method}

\item{params}{the vector of parameters for which we are computing the likelihood}
}
\description{
Estimates the rate parameters for a general multitype branching process with time-dependent rates
}
