% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimateBP}
\alias{estimateBP}
\title{estimateBP}
\usage{
estimateBP(time, N, transitionList, data, initial, known = NULL)
}
\arguments{
\item{time}{numeric or vector of time units for data observations}

\item{N}{intial state vector}

\item{transitionList}{TransitionList object specifying transitions in system}

\item{data}{n x k data matrix}

\item{initial}{vector of initial estimates for MLE optimization}

\item{known}{boolean vector of known parameter rates, if NULL, all rates will be estimated}
}
\description{
Estimates the rate parameters for a general multitype branching process
}
