% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimateBP_timedep}
\alias{estimateBP_timedep}
\title{estimateBP_timedep}
\usage{
estimateBP_timedep(time, N, initTime, transitionList, data, initial,
  known = NULL, lower = NULL, upper = NULL, trace = 1)
}
\arguments{
\item{time}{numeric or vector of time units for data observations}

\item{N}{intial state vector}

\item{initTime}{time of observation for initial state vector}

\item{transitionList}{TransitionList object specifying transitions in system}

\item{data}{n x k data matrix}

\item{initial}{vector of initial estimates for MLE optimization}

\item{known}{boolean vector of known parameter rates, if NULL, all rates will be estimated}

\item{lower}{vector of lower bounds on rate parameters for optimization}

\item{upper}{vector of upper bounds on rate parameters for optimization}

\item{trace}{level of output for optimizer - see optim function, control - trace for "L-BFGS-B" method}
}
\description{
Estimates the rate parameters for a general multitype branching process with time-dependent rates
}
