% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmbp.R
\docType{package}
\name{estipop}
\alias{estipop}
\alias{estipop-package}
\title{estipop: simualtion and estimation of general multitype branching processes}
\description{
estipop: simualtion and estimation of general multitype branching processes
}
