% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{FixedTransition}
\alias{FixedTransition}
\title{FixedTransition}
\usage{
FixedTransition(population, rate, fixed)
}
\arguments{
\item{population}{index of population for which transition applies}

\item{rate}{rate at which this transition occurs}

\item{fixed}{}
}
\description{
Designates a fixed transition, that is, a transition vector that does not contain random elements, such as a split event
}
\examples{
\dontrun{
FixedTransition(population = 0, rate = 0.5, fixed = c(1, 1, 0))
}
}
