% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generateCpp}
\alias{generateCpp}
\title{generateCPP}
\usage{
generateCpp(exprn, params)
}
\arguments{
\item{exprn}{the R expression to be translated}

\item{params}{the vectors of parameter to substitute into the expression}
}
\description{
Takes an expression and generates CPP code to compute it
}
