% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/gmbp.R
\docType{package}
\name{gmbp}
\alias{gmbp}
\alias{gmbp}
\alias{gmbp-package}
\title{gmbp}
\usage{
gmbp(time, file, initial, lifetimes, transitions)
}
\description{
gmbp

gmbp: simualtion of tree-like structures
}
