% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_const}
\alias{is_const}
\title{is_const
 
helper for determining whether an expression depends on time}
\usage{
is_const(ast)
}
\description{
is_const
 
helper for determining whether an expression depends on time
}
