% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{loglik_est_time}
\alias{loglik_est_time}
\title{loglik_est_time}
\usage{
loglik_est_time(dat, t, N, parent, rate, offspring)
}
\arguments{
\item{dat}{Nxk matrix of observed data where each column is a count from a type and each row is an observation}

\item{t}{N-length vecotr of time from system initialization until each observation in dat were made}

\item{N}{k-length vector of the initial ancestor counts for each type}

\item{parent}{a d-length vector specifying which population type is associated with a specific offspring transition}

\item{rate}{a d-length vector specifying the rate at which each offspring transition is occurring}

\item{offspring}{a dxk matrix specifying the offspring transitions}
}
\description{
log-likelihood function for N replicates various time observations from a general k-type branching process with d transitions
}
