% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{loglik_time_dependent}
\alias{loglik_time_dependent}
\title{loglik_time_dependent}
\usage{
loglik_time_dependent(dat, t, N, parent, rate_func, rate_params, offspring)
}
\arguments{
\item{dat}{Nxk matrix of observed data where each column is a count from a type and each row is an observation}

\item{t}{N-length vector of time from system initialization until each observation in dat were made}

\item{N}{Nxk matrix of initial ancestor counts for each type for each observation}

\item{parent}{a d-length vector specifying which population type is associated with a specific offspring transition}

\item{rate_func}{a function of time returning a d-length vector specifying the rate at which each offspring transition is occurring}

\item{rate_params}{a vector of parameters for the rate functions. These are the parameters than will be estimated}

\item{offspring}{a dxk matrix specifying the offspring transitions}
}
\description{
log-likelihood function for N replicates from various time observations and possibly different initial counts from a general time-inhomogenous k-type branching process with d transitions
}
