% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{moments}
\alias{moments}
\title{moments}
\usage{
moments(model, params, start_times, end_times)
}
\arguments{
\item{model}{the \code{process_model} object representing the process whose moments will be computed}

\item{params}{the vector of parameters to plug into the process model during moment computation}

\item{start_times}{the various start times of the moments to be computed}

\item{end_times}{the various end times of the moments to be computed}
}
\value{
a dataframe with the moments vector for each unique start and end combination
}
\description{
compute the moment matrices for a time-inhomogenous branching process model with a certain set of parameters. Helps for computing likelihoods.
}
