% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{new_process_model}
\alias{new_process_model}
\title{new_process_model
constructor for object of class \code{process_model}}
\usage{
new_process_model(transition_list)
}
\arguments{
\item{transition_list}{a list of estipop transition objects}
}
\value{
A new branching process model object
}
\description{
new_process_model
constructor for object of class \code{process_model}
}
