% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{new_rate}
\alias{new_rate}
\title{new_rate
constructor for object of class \code{rate}}
\usage{
new_rate(exprn)
}
\arguments{
\item{exprn}{an expression dictating the functional dependence of the rate on parameters and on time}
}
\value{
A new branching process rate object
}
\description{
new_rate
constructor for object of class \code{rate}
}
