% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{new_stop_criterion}
\alias{new_stop_criterion}
\title{new_stop_critereon}
\usage{
new_stop_criterion(indices, inequality, value)
}
\arguments{
\item{indices}{the indices of population vector to sum in determining whether to stop}

\item{inequality}{the type of comparison}

\item{value}{value to compare sum against to determine whether to stop simulation}
}
\description{
constructor for class of type \code{stop_criteron}
}
