% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{new_stoplist}
\alias{new_stoplist}
\title{new_stoplist}
\usage{
new_stoplist(stops)
}
\arguments{
\item{stops}{the list of \code{stop_criteron} objects to add to the stoplist object}
}
\value{
new \code{stop_list} object
}
\description{
constructor for class of type \code{stop_list}
}
