% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{new_transition}
\alias{new_transition}
\title{new_transition
constructor for object of class \code{transition}}
\usage{
new_transition(rate, parent, offspring)
}
\arguments{
\item{rate}{a rate for the transition}

\item{parent}{the parent type for the transition}

\item{offspring}{the update to the system after the transition}
}
\value{
A new branching process rate object
}
\description{
new_transition
constructor for object of class \code{transition}
}
