% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{process_model}
\alias{process_model}
\title{process_model
constructs and validates an object of class \code{process_model}}
\usage{
process_model(...)
}
\arguments{
\item{transitions}{a list of estipop transition objects}
}
\value{
The \code{process_model} object if it is valid, throws an error otherwise
}
\description{
process_model
constructs and validates an object of class \code{process_model}
}
