% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RandomTransition}
\alias{RandomTransition}
\title{RandomTransition}
\usage{
RandomTransition(population, rate, oVec, oDist, oParams)
}
\arguments{
\item{population}{index of population for which transition applies}

\item{rate}{rate at which this transition occurs}

\item{oVec}{offspring distribution probabilities}

\item{oDist}{offspring distribution function}

\item{oParams}{offspring distriution function parameters}
}
\description{
Designates a fixed transition, that is, a transition vector that generates the number of offspring according to the distribution oDist with
parameters oParams, and then divides these according to a multinomial distribution with probability vector oVec
}
\examples{
\dontrun{
RandomTransition(population = 0, rate = 0.5, fixed = c(1, 1, 0))
}
}
