% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{rate}
\alias{rate}
\title{rate
constructs and validates an object of class \code{rate}}
\usage{
rate(exprn)
}
\arguments{
\item{exprn}{an expression definintion a functional dependence of rate on parameters and on time}
}
\value{
The \code{rate} object if it is valid, throws an error otherwise
}
\description{
rate
constructs and validates an object of class \code{rate}
}
