% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reload}
\alias{reload}
\title{Utility functions for time-dependent rates
 taken with minor adaptions from:}
\usage{
reload(path)
}
\arguments{
\item{path}{path to siapopr package source}
}
\description{
McDonald, T. O., & Michor, F. (2017).
 SIApopr: a computational method to simulate evolutionary branching trees for
 analysis of tumor clonal evolution. Bioinformatics, 33(14), 2221-2223.
}
\details{
Original code author: T. O. McDonald (mcdonald(at)jimmy.harvard.edu)

reload

Recompiles and reinstalls the package for use after adjusting the model
code. WARNING: For advanced users. Adjusting the C++ code can lead to
errors in compilation.
}
