% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_approx.R
\name{sim_approx_full}
\alias{sim_approx_full}
\title{sim_approx_full}
\usage{
sim_approx_full(numSamples, t, N, transitionList)
}
\arguments{
\item{numSamples}{number of samples}

\item{t}{time to simulate until}

\item{N}{k-length vector of the initial ancestor counts for each type}

\item{transitionList}{transitionList object to specify the model}
}
\description{
method to simulate from the asymptotic distriution of a general multitype branching process
}
