% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyFunctions.R
\name{simBirth}
\alias{simBirth}
\title{simBirth}
\usage{
simBirth(birth = 1, init = 100, time = 1)
}
\arguments{
\item{birth}{constant birth rate}

\item{init}{initial population size}

\item{time}{number of time units to simulate}
}
\description{
Simulates a 1-type linear birth process
}
\examples{
\dontrun{
simBirth(birth = 1, init = 100, time = 10)
}
}
