% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyFunctions.R
\name{simBirthDeath}
\alias{simBirthDeath}
\title{simBirthDeath}
\usage{
simBirthDeath(birth = 1, death = 0, init = 100, time = 1)
}
\arguments{
\item{birth}{constant birth rate}

\item{death}{constant death rate}

\item{init}{initial population size}

\item{time}{number of time units to simulate}
}
\description{
Simulates a 1-type linear birth-death process
}
\examples{
\dontrun{
simBirthDeath(birth = 1, death = 0.7, init = 100, time = 10)
}
}
