% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyFunctions.R
\name{simBirthMutation}
\alias{simBirthMutation}
\title{simBirthMutation}
\usage{
simBirthMutation(birth1 = 1, mutation = 0, birth2 = 1, init = c(100, 0),
  time = 1, approx = FALSE, dtime = NULL, observations = NULL)
}
\arguments{
\item{birth1}{constant birth rate of population 1}

\item{mutation}{constant mutation rate}

\item{birth2}{constant birth rate of population 2}

\item{init}{vector of initial population sizes}

\item{time}{number of time units to simulate}

\item{approx}{boolean if TRUE, approximate simulation is used, if FALSE, exact simulation is performed}

\item{dtime}{if dtime and time are not null, observations will be made from time 0 to time in units of dtime}

\item{observations}{numeric vector specifying custom observation times}
}
\description{
Simulates a 2-type birth-mutation process
}
\examples{
\dontrun{
simBirthMutation(birth1 = 1, death1 = 0.7, mutation = 0.1, birth2 = 1.0, death2 = 0.7, init = c(100,0), time = 10)
}
}
