% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyFunctions.R
\name{simpBirthDeath}
\alias{simpBirthDeath}
\title{simpBirthDeath}
\usage{
simpBirthDeath(b = 1, d = 0, init = 100, t = 1, N = 1)
}
\arguments{
\item{b}{constant birth rate}

\item{d}{constant death rate}

\item{init}{initial population size}

\item{t}{simulate at which time}

\item{N}{how many samples}
}
\description{
Simulates a 1-type linear birth-death process
}
\examples{
\dontrun{
simpBirthDeath(birth = 1, death = 0.7, init = 100, time = 10)
}
}
