% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyFunctions.R
\name{simTwoTypeResistance}
\alias{simTwoTypeResistance}
\title{simTwoTypeResistance}
\usage{
simTwoTypeResistance(birth1 = 1, death1 = 0, mutation = 0, birth2 = 1,
  death2 = 0.7, init = c(100, 0), time = 1)
}
\arguments{
\item{birth1}{constant birth rate of population 1}

\item{death1}{constant death rate of population 1}

\item{birth2}{constant birth rate of population 2}

\item{death2}{constant death rate of population 2}

\item{init}{vector of initial population sizes}

\item{time}{number of time units to simulate}

\item{birth0}{constant birth rate of population 0}

\item{death0}{constant death rate of population 0}

\item{mutation1}{constant mutation rate to type-1 resistance}

\item{mutation2}{constant mutation rate to type-2 resistance}

\item{mutation12}{constant mutation rate from type-1 resistance to include type-2 resistance}

\item{mutation21}{constant mutation rate from type-2 restistance to include type-1 resistance}

\item{birth12}{constant birth rate of population 1/2}

\item{death12}{constant death rate of population 1/2}
}
\description{
Simulates 2-type Resistance process
}
\examples{
\dontrun{
simBirthDeathMutation(birth1 = 1, death1 = 0.7, mutation = 0.1, birth2 = 1.0, death2 = 0.7, init = c(100,0), time = 10)
}
}
