% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{stop_criterion}
\alias{stop_criterion}
\title{stop_criteron
constructs and validates an object of class \code{stop_critereon}}
\usage{
stop_criterion(indices, inequality, value)
}
\arguments{
\item{indices}{the indices of population vector to sum in determining whether to stop}

\item{inequality}{the type of comparison}

\item{value}{value to compare sum against to determine whether to stop simulation}
}
\value{
the \code{stop_criteron} object if it is valid, throws an error otherwise
}
\description{
stop_criteron
constructs and validates an object of class \code{stop_critereon}
}
