% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{StopCriterion}
\alias{StopCriterion}
\title{StopCriterion}
\usage{
StopCriterion(indices, inequality, value)
}
\arguments{
\item{indices}{indices of states to be included in the sum}

\item{inequality}{">", ">=}

\item{value}{value to compare the sum of states to}
}
\description{
Designates a condition upon which the simulation should stop
}
\details{
If the sum of the specified states meets the inequality specified by the particular inequality and value, the simulation stops
}
\examples{
\dontrun{
StopCriterion(indices = c(0), inequality = ">=", value = 1000)
}
}
