% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{StopList}
\alias{StopList}
\title{StopList}
\usage{
StopList(...)
}
\arguments{
\item{...}{StopCriterion objects}
}
\description{
Makes a single StopList object out of multiple StopCriterion objects
}
\examples{
\dontrun{
StopList(StopCriterion(indices = c(0), inequality = ">=", value = 1000),
         StopCriterion(indices = c(0, 1), inequality = ">=", value = 10000))
}
}
