% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Transition}
\alias{Transition}
\title{Transition}
\usage{
Transition(population, prob, fixed, random = NULL)
}
\arguments{
\item{population}{index of population for which transition applies}

\item{prob}{probability that this transition occurs}

\item{random}{}
}
\description{
Designates a Transition, that is, a transition vector that may contain or may not contain random elements
}
\examples{
\dontrun{
Transition(population = 0, prob = 0.5, fixed = c(1, 1, 0), random = c(FALSE, FALSE, TRUE))
}
}
