% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{TransitionList}
\alias{TransitionList}
\title{TransitionList}
\usage{
TransitionList(...)
}
\arguments{
\item{...}{Transition objects}
}
\description{
Makes a single TransitionList object out of multiple Transition objects
}
\examples{
\dontrun{
TransitionList(Transition(prob = 0.5, fixed = c(1, 1, 0), random = c(FALSE, FALSE, TRUE)),
               Transition(prob = 0.5, fixed = c(0, 0, 0), random = c(TRUE, FALSE, FALSE)))
}
}
