% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{validate_rate}
\alias{validate_rate}
\title{validate_rate
verifies the correctness of an object of class \code{rate}}
\usage{
validate_rate(rate_obj)
}
\value{
The \code{rate} object if it is valid, throws an error otherwise
}
\description{
validate_rate
verifies the correctness of an object of class \code{rate}
}
