% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{validate_stop_criteron}
\alias{validate_stop_criteron}
\title{validate_stop_criterion}
\usage{
validate_stop_criteron(sc_obj)
}
\arguments{
\item{the}{\code{stop_criteron} object to validate}
}
\value{
The \code{stop_criteron} object if it is valid, throws an error otherwise
}
\description{
verifies the correctness of a \code{stop_critereon} object
}
