% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{validate_stoplist}
\alias{validate_stoplist}
\title{validate_stop_criterion}
\usage{
validate_stoplist(sl)
}
\arguments{
\item{sl}{the \code{stop_list} object to validate}
}
\value{
The \code{stop_list} object if it is valid, throws an error otherwise
}
\description{
verifies the correctness of a \code{stop_list} object
}
