% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{validate_transition}
\alias{validate_transition}
\title{validate_transition
verifies the correctness of an object of class \code{transition}}
\usage{
validate_transition(trans_obj)
}
\value{
The \code{transition} object if it is valid, throws an error otherwise
}
\description{
validate_transition
verifies the correctness of an object of class \code{transition}
}
