% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{bploglikelihood}
\alias{bploglikelihood}
\title{bploglikelihood}
\usage{
bploglikelihood(dat, t, N, transitionList)
}
\arguments{
\item{dat}{nxK matrix of observed data where each column is a count from a type and each row is an observation}

\item{t}{time from system initialization until observations in dat were made}

\item{N}{initial ancestor counts for each type}

\item{transitionList}{A TransitionList object to specify the model form}
}
\description{
log-likelihood function for N replicates from a general k-type branching process with d transitions
}
