% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{branch}
\alias{branch}
\title{branch}
\usage{
branch(time, initial, transitionList, stopList, silent = FALSE,
  keep = FALSE, seed = NULL)
}
\arguments{
\item{time}{number of time units to simulate}

\item{initial}{intial state vector}

\item{transitionList}{TransitionList object specifying transitions in system}

\item{stopList}{StopList object specifying stopping conditions for the system}

\item{silent}{if true, verbose output will be shown.  Default: false}

\item{keep}{if true, the temporary comma-separated file generated while simulating while be kept.  if false, it will be deleted.  Default: false}

\item{seed}{seed for the random number generator.  If NULL, will use computer clock to set a random seed}
}
\description{
Makes a single StopList object out of multiple StopCriterion objects
}
\examples{
\dontrun{
branch(time = 100,
       initial = c(1,0),
       transistionList = TransitionList(Transition(prob = 0.5, fixed = c(1, 1, 0), random = c(FALSE, FALSE, TRUE)),
                                        Transition(prob = 0.5, fixed = c(0, 0, 0), random = c(TRUE, FALSE, FALSE))),
       stopList = StopList(StopCriterion(indices = c(0), inequality = ">=", value = 1000),
                  StopCriterion(indices = c(0, 1), inequality = ">=", value = 10000)))
}
}
