% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{branch}
\alias{branch}
\title{branch}
\usage{
branch(time, intial, transitionList, stopList)
}
\arguments{
\item{time}{number of time units to simulate}

\item{transitionList}{TransitionList object specifying transitions in system}

\item{stopList}{StopList object specifying stopping conditions for the system}

\item{initial}{intial state vector}
}
\description{
Makes a single StopList object out of multiple StopCriterion objects
}
\examples{
\dontrun{
branch(time = 100,
       initial = c(1,0),
       transistionList = TransitionList(Transition(prob = 0.5, fixed = c(1, 1, 0), random = c(FALSE, FALSE, TRUE)),
                                        Transition(prob = 0.5, fixed = c(0, 0, 0), random = c(TRUE, FALSE, FALSE))),
       stopList = StopList(StopCriterion(indices = c(0), inequality = ">=", value = 1000),
                  StopCriterion(indices = c(0, 1), inequality = ">=", value = 10000)))
}
}
