% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{estimateBP}
\alias{estimateBP}
\title{estimateBP}
\usage{
estimateBP(time, N, transitionList, data, initial, ... = NULL)
}
\arguments{
\item{time}{numeric or vector of time units for data observations}

\item{N}{intial state vector}

\item{transitionList}{TransitionList object specifying transitions in system}

\item{data}{n x k data matrix}

\item{initial}{vector of initial estimates for MLE optimization}

\item{...}{additional parameters to pass to optimizer}
}
\description{
Estimates the rate parameters for a general multitype branching process
}
\examples{
\dontrun{
estimateBP(time = 100,
       initial = c(1,0),
       transistionList = TransitionList(Transition(prob = 0.5, fixed = c(1, 1, 0), random = c(FALSE, FALSE, TRUE)),
                                        Transition(prob = 0.5, fixed = c(0, 0, 0), random = c(TRUE, FALSE, FALSE))),
       stopList = StopList(StopCriterion(indices = c(0), inequality = ">=", value = 1000),
                  StopCriterion(indices = c(0, 1), inequality = ">=", value = 10000)))
}
}
