% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Rate}
\alias{Rate}
\title{Rate}
\usage{
Rate(expr)
}
\arguments{
\item{expr}{the expression representing the function of time}
}
\description{
Specifies a rate.  User writes an R expression which can involve time (t) or params[1..n]
}
\examples{
\dontrun{
Rate(params[0]*t + params[1])
}

}
