% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Rate}
\alias{Rate}
\title{Rate}
\usage{
Rate(type, params)
}
\arguments{
\item{type}{type of rate function, see README}

\item{params}{list of parameters}
}
\description{
Designates a rate
}
\details{
Specifies a rate.  User can specify using built-in types and accompanying parameters, or use a custom C++ function of time
}
\examples{
\dontrun{
Rate(type = 0, params = c(0.01))
Rate(type = 1, params = c(0.1, -0.01))
}
}
