% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_approx.R
\name{sim_approx}
\alias{sim_approx}
\title{sim_approx}
\usage{
sim_approx(numSamples, t, N, transitionList)
}
\arguments{
\item{numSamples}{number of samples}

\item{t}{N-length vector of time from system initialization until each observation in dat were made}

\item{N}{k-length vector of the initial ancestor counts for each type}

\item{transitionList}{transitionList object to specify the model}
}
\description{
method to simulate from the asymptotic distriution of a general multitype branching process
}
