% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{stoplist}
\alias{stoplist}
\title{stoplist}
\usage{
stoplist(...)
}
\arguments{
\item{...}{a series of \code{stop_criteron} objects to add to the stoplist object}
}
\value{
new \code{stop_list} object if valid, throws an error otherwise
}
\description{
constucts and validates an object of type \code{stop_list}
}
