% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{transition}
\alias{transition}
\title{transition 
constructs and validates an object of class \code{transition}}
\usage{
transition(rate, parent, offspring)
}
\arguments{
\item{rate}{a rate for the transition}

\item{parent}{the parent type for the transition}

\item{offspring}{the update to the system after the transition}
}
\value{
The \code{transition} object if it is valid, throws an error otherwise
}
\description{
transition 
constructs and validates an object of class \code{transition}
}
